refactor: separate backgroundImg to a component
